+++ src/lib/controllers/editor/index.ts
@@ -1,2 +1,0 @@
-export * from './createEditorShortcuts.svelte';
-export * from './createEditorPaneController.svelte';