+++ src/lib/controllers/index.ts
@@ -1,8 +1,9 @@
 export * from './activity/index';
-export * from './editor/index';
 export * from './explorer/index';
 export * from './workspace/index';
 
 export * from './LiveblocksSyncController.svelte';
 export * from './FileTreeController.svelte';
 export * from './HeaderController.svelte';
+export * from './EditorController.svelte';
+export * from './ShortcutsController.svelte';