+++ src/routes/(app)/+layout.svelte
@@ -4,17 +4,6 @@
 	let { children } = $props();
 </script>
 
-<div class="container">
-	<AppHeader />
+<AppHeader />
 
-	{@render children()}
-</div>
-
-<style>
-	.container {
-		height: 100dvh;
-		overflow-y: auto;
-		overflow-x: hidden;
-		scrollbar-gutter: stable;
-	}
-</style>
+{@render children()}