+++ src/routes/(app)/auth/+page.svelte
@@ -3,7 +3,6 @@
 	import { api } from '$convex/_generated/api.js';
 	import { useQuery } from 'convex-svelte';
 	import { useAuth } from '$lib/svelte/index.js';
-	import type { ErrorContext } from '@better-fetch/fetch';
 	import Button from '$lib/components/ui/primitives/Button.svelte';
 	import Form from '$lib/components/ui/primitives/Form.svelte';
 	import Grid from '$lib/components/ui/primitives/Grid.svelte';
@@ -39,8 +38,8 @@
 				await authClient.signIn.email(
 					{ email, password },
 					{
-						onError: (ctx: ErrorContext) => {
-							errorMsg = ctx.error?.message ?? 'Unable to sign in';
+						onError: (ctx) => {
+							errorMsg = ctx.error.message;
 						}
 					}
 				);
@@ -48,8 +47,8 @@
 				await authClient.signUp.email(
 					{ name, email, password },
 					{
-						onError: (ctx: ErrorContext) => {
-							errorMsg = ctx.error?.message ?? 'Unable to sign up';
+						onError: (ctx) => {
+							errorMsg = ctx.error.message;
 						}
 					}
 				);